refactor(message-list): revert the `BaseMessageListFragment` and duplicate the `LegacyMessageListFragment` within the `MessageListFragment` (#10559)
