+++ app-k9mail/src/debug/kotlin/app/k9mail/featureflag/K9FeatureFlagFactory.kt
@@ -25,7 +25,7 @@ class K9FeatureFlagFactory : FeatureFlagFactory {
                 FeatureFlag(AccountSettingsFeatureFlags.EnableAvatarCustomization, enabled = true),
                 // TODO(#10498): Remove when UseNewMessageReaderCssStyles is no longer required
                 FeatureFlag(MessageReaderFeatureFlags.UseNewMessageReaderCssStyles, enabled = true),
-                FeatureFlag(MessageListFeatureFlags.EnableMessageListNewState, enabled = true),
+                FeatureFlag(MessageListFeatureFlags.EnableMessageListNewState, enabled = false),
             ),
         )
     }
