+++ legacy/ui/legacy/src/main/java/com/fsck/k9/ui/KoinModule.kt
@@ -7,9 +7,9 @@ import app.k9mail.legacy.message.controller.MessagingControllerMailChecker
 import com.fsck.k9.controller.MessagingController
 import com.fsck.k9.ui.helper.DisplayHtmlUiFactory
 import com.fsck.k9.ui.helper.SizeFormatter
-import com.fsck.k9.ui.messagelist.BaseMessageListFragment
 import com.fsck.k9.ui.messagelist.LegacyMessageListFragment
 import com.fsck.k9.ui.messagelist.MessageListFragment
+import com.fsck.k9.ui.messagelist.MessageListFragmentBridgeContract
 import com.fsck.k9.ui.messageview.LinkTextHandler
 import com.fsck.k9.ui.settings.AboutViewModel
 import com.fsck.k9.ui.share.ShareIntentBuilder
@@ -38,7 +38,7 @@ val uiModule = module {
     factory { (activity: Activity) ->
         FoldableStateObserver(activity = activity, logger = get())
     }
-    factory<BaseMessageListFragment.Factory> {
+    factory<MessageListFragmentBridgeContract.Factory> {
         val featureFlagProvider = get<FeatureFlagProvider>()
         if (featureFlagProvider.provide(MessageListFeatureFlags.EnableMessageListNewState).isEnabled()) {
             MessageListFragment.Factory
