+++ legacy/ui/legacy/src/main/res/layout/message_list_fragment.xml
@@ -6,7 +6,7 @@
     android:id="@+id/message_list_coordinator"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".messagelist.BaseMessageListFragment"
+    tools:context=".messagelist.LegacyMessageListFragment"
     >
     <com.google.android.material.floatingactionbutton.FloatingActionButton
         android:id="@+id/floating_action_button"
